requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pinecone-client>=3.0.0
python-multipart>=0.0.5,<0.0.6
python-dotenv>=0.19.0,<0.20.0
httpx>=0.23.0,<0.24.0
